requirements.txt
# Python >= 3.10

# -----------------------------
# MCP servers (MCPServers/*.py)
# -----------------------------
fastmcp
mcp>=1.5.0
requests>=2,<3
lief
keystone-engine

# FLOSS/capa command dependencies used by MCP tools
flare-floss
flare-capa

# -----------------------------
# Client (multi_agent_wf/main.py)
# -----------------------------
gradio
pydantic
pydantic-ai
pydantic-ai-backend
python-dotenv

# Required deep-agent orchestration support in multi_agent_wf/main.py
pydantic-deep

# Optional alternate client in In_Progress/multiClient.py
langchain
langchain-openai
langchain-mcp-adapters

# Optional in-progress Ghidra scripting server
pyghidra

# Optional helper dependencies for bundled/vendored maintenance scripts
PyYAML
PyGithub

# Testing infrastructure, launch doctor, and sweep visualization
# (Testing/scripts/run_evaluation.py, Testing/scripts/run_experiment_sweep.py,
#  Testing/scripts/run_launch_doctor.py, and Testing/harness/*)
pandas>=1.5
matplotlib>=3.7
openpyxl>=3.1
